Organize Harrier clustering like v0
