+++ experiments/datakit/cluster/domain/v1/harrier_all_sources.py
@@ -3,39 +3,18 @@
 
 """Train Harrier domain clusters on a proportional sample of every source."""
 
-import hashlib
-import json
 import logging
-import time
-from collections.abc import Iterator
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass
-from typing import Any, Protocol
 
-import numpy as np
-import pyarrow as pa
-import pyarrow.parquet as pq
 from fray.types import ResourceConfig
-from marin.datakit.source_key import datakit_source_path
 from marin.execution.remote import remote
 from marin.execution.step_runner import StepRunner
 from marin.execution.step_spec import StepSpec
-from rigging.filesystem import StoragePath, open_url, prefix_join
 from rigging.log_setup import configure_logging
-from zephyr import counters
-from zephyr.dataset import Dataset, ShardInfo
-from zephyr.execution import ZephyrContext
-from zephyr.readers import InputFileSpec, load_file
-from zephyr.runners import InlineRunner
 
-from experiments.datakit.cluster.domain.v1.coarsen import (
-    ClusteringStats,
-    CoarseningConfig,
-    CoarseningResult,
-    coarsen_centroids,
-)
-from experiments.datakit.embeddings.harrier.pipeline import HARRIER_DIM, dequantize_to_fp32
-from experiments.datakit.reference_pipeline import select_sources
+from experiments.datakit.cluster.domain.v1.coarsen import CoarseningConfig
+from experiments.datakit.cluster.domain.v1.sample import sample_centroid_inputs
+from experiments.datakit.cluster.domain.v1.train import train_centroids
+from experiments.datakit.embeddings.harrier.run import build_steps as build_harrier_embeddings
 
 logger = logging.getLogger(__name__)
 
@@ -54,7 +33,7 @@ PARALLEL_SOURCES = 16
 MAX_WORKERS = 64
 LOAD_PARALLELISM = 64
 ASSIGN_BATCH_ROWS = 65_536
-EMBEDDING_ROOT = datakit_source_path("datakit/embed/harrier")
+PIPELINE_VERSION = "2026.08.13"
 
 COARSENING_CONFIG = CoarseningConfig(
     clusters=K_COARSE,
@@ -63,6 +42,11 @@ COARSENING_CONFIG = CoarseningConfig(
     split_restarts=10,
     split_iterations=20,
 )
+
+DRIVER_RESOURCES = ResourceConfig.with_cpu(cpu=2, ram="16g", preemptible=False)
+SAMPLE_WORKER_RESOURCES = ResourceConfig.with_cpu(cpu=2, ram="16g")
+SAMPLE_COORDINATOR_RESOURCES = ResourceConfig.with_cpu(cpu=1, ram="8g", preemptible=False)
+TRAIN_RESOURCES = ResourceConfig.with_cpu(cpu=N_THREADS, ram="256g", disk="32g", preemptible=False)
 
 
 def _coarsening_cache_key(config: CoarseningConfig) -> str:
@@ -76,440 +60,34 @@ def _coarsening_cache_key(config: CoarseningConfig) -> str:
 
 COARSENING_CACHE_KEY = _coarsening_cache_key(COARSENING_CONFIG)
 
-DRIVER_RESOURCES = ResourceConfig.with_cpu(cpu=2, ram="16g", preemptible=False)
-SAMPLE_WORKER_RESOURCES = ResourceConfig.with_cpu(cpu=2, ram="16g")
-TRAIN_RESOURCES = ResourceConfig.with_cpu(cpu=N_THREADS, ram="256g", disk="32g", preemptible=False)
-
-_SAMPLE_SCHEMA = pa.schema(
-    [
-        pa.field("source", pa.string()),
-        pa.field("embedding", pa.list_(pa.int8(), HARRIER_DIM)),
-    ]
-)
-
-
-@dataclass(frozen=True)
-class SourceSample:
-    source: str
-    paths: tuple[str, ...]
-    rows: tuple[int, ...]
-    quota: int
-
-
-@dataclass(frozen=True)
-class _TrainingData:
-    embeddings: np.ndarray
-    population: np.ndarray
-    rows_by_source: dict[str, int]
-
-
-class _SearchIndex(Protocol):
-    def search(self, values: np.ndarray, k: int) -> tuple[np.ndarray, np.ndarray]: ...
-
-
-@dataclass(frozen=True)
-class _TrainedCentroids:
-    centroids: np.ndarray
-    index: _SearchIndex
-    seconds: float
-    objective: float
-
-
-def _largest_remainder_quotas(weights: tuple[int, ...], target: int) -> tuple[int, ...]:
-    total = sum(weights)
-    if target == 0:
-        return (0,) * len(weights)
-    if total == 0:
-        raise ValueError("cannot allocate a positive quota across zero available rows")
-    numerators = [target * weight for weight in weights]
-    quotas = [numerator // total for numerator in numerators]
-    leftover = target - sum(quotas)
-    order = sorted(range(len(weights)), key=lambda index: (-(numerators[index] % total), index))
-    for index in order[:leftover]:
-        quotas[index] += 1
-    return tuple(quotas)
-
-
-def proportional_quotas(counts: dict[str, int], target: int) -> dict[str, int]:
-    if target < len(counts) * SMALL_SOURCE_QUOTA:
-        raise ValueError("target is too small for the one-document source floor")
-    if target > sum(counts.values()):
-        raise ValueError("target exceeds the available documents")
-
-    small = {source for source, rows in counts.items() if rows <= SMALL_SOURCE_MAX_ROWS}
-    quotas = {source: SMALL_SOURCE_QUOTA for source in small}
-    large_sources = tuple(sorted(source for source in counts if source not in small))
-    remaining = target - sum(quotas.values())
-    large_quotas = _largest_remainder_quotas(tuple(counts[source] for source in large_sources), remaining)
-    quotas.update(dict(zip(large_sources, large_quotas, strict=True)))
-    if any(quota > counts[source] for source, quota in quotas.items()):
-        raise ValueError("a source quota exceeds its document count")
-    return quotas
-
-
-def training_quotas(counts: dict[str, int], target: int) -> dict[str, int]:
-    if target < len(counts):
-        raise ValueError("target is too small for the one-document source floor")
-    if target > sum(counts.values()):
-        raise ValueError("target exceeds the available documents")
-
-    sources = tuple(sorted(counts))
-    remaining = target - len(counts)
-    if remaining == 0:
-        return {source: 1 for source in sources}
-    extra = _largest_remainder_quotas(tuple(counts[source] - 1 for source in sources), remaining)
-    return {source: quota + 1 for source, quota in zip(sources, extra, strict=True)}
-
-
-def _stratified_indices(
-    source_codes: np.ndarray,
-    source_names: tuple[str, ...],
-    target: int,
-    seed: int,
-) -> tuple[np.ndarray, dict[str, int]]:
-    counts_array = np.bincount(source_codes, minlength=len(source_names))
-    counts = dict(zip(source_names, counts_array.tolist(), strict=True))
-    quotas = training_quotas(counts, target)
-    selected = []
-    source_codes_by_name = {source: code for code, source in enumerate(source_names)}
-    for source in sorted(source_names):
-        code = source_codes_by_name[source]
-        source_indices = np.flatnonzero(source_codes == code)
-        quota = quotas[source]
-        if quota == len(source_indices):
-            selected.append(source_indices)
-            continue
-        digest = hashlib.sha256(f"{seed}:{source}".encode()).digest()
-        rng = np.random.default_rng(int.from_bytes(digest[:8], "little"))
-        selected.append(rng.choice(source_indices, size=quota, replace=False))
-    return np.sort(np.concatenate(selected)), quotas
-
-
-def _part_quotas(rows: tuple[int, ...], target: int) -> tuple[int, ...]:
-    return _largest_remainder_quotas(rows, target)
-
-
-def _row_count(path: str) -> int:
-    with StoragePath(path).open("rb") as file:
-        return pq.ParquetFile(file).metadata.num_rows
-
-
-def _source_samples(embedding_paths: dict[str, str]) -> tuple[SourceSample, ...]:
-    source_paths = {
-        source: tuple(sorted(str(shard) for shard in StoragePath(prefix_join(path, "*.parquet")).glob()))
-        for source, path in sorted(embedding_paths.items())
-    }
-    if any(not paths for paths in source_paths.values()):
-        missing = [source for source, paths in source_paths.items() if not paths]
-        raise ValueError(f"sources have no embedding shards: {missing}")
-
-    all_paths = [path for paths in source_paths.values() for path in paths]
-    with ThreadPoolExecutor(max_workers=LOAD_PARALLELISM) as pool:
-        all_rows = list(pool.map(_row_count, all_paths))
-    path_rows = dict(zip(all_paths, all_rows, strict=True))
-    source_rows = {source: sum(path_rows[path] for path in paths) for source, paths in source_paths.items()}
-    empty_sources = [source for source, rows in source_rows.items() if rows == 0]
-    if empty_sources:
-        logger.warning("Skipping %d zero-row sources: %s", len(empty_sources), empty_sources)
-        source_paths = {source: paths for source, paths in source_paths.items() if source_rows[source]}
-        source_rows = {source: rows for source, rows in source_rows.items() if rows}
-    quotas = proportional_quotas(source_rows, TARGET_ROWS)
-    return tuple(
-        SourceSample(
-            source=source,
-            paths=paths,
-            rows=tuple(path_rows[path] for path in paths),
-            quota=quotas[source],
-        )
-        for source, paths in sorted(source_paths.items())
-    )
-
-
-def _completed_embedding_paths(source_names: tuple[str, ...]) -> dict[str, str]:
-    embedding_root = StoragePath(EMBEDDING_ROOT)
-    records = [
-        StoragePath(path)
-        for path in sorted(str(path) for path in StoragePath(prefix_join(EMBEDDING_ROOT, "**/.artifact.json")).glob())
-    ]
-    candidates: dict[str, list[str]] = {source: [] for source in source_names}
-    source_prefixes = sorted(source_names, key=len, reverse=True)
-    for record in records:
-        output_path = record.parent
-        relative = output_path.relative_to(embedding_root)
-        for source in source_prefixes:
-            if relative.startswith(f"{source}_"):
-                candidates[source].append(str(output_path))
-                break
-
-    resolved = {}
-    for source, paths in candidates.items():
-        if not paths:
-            raise ValueError(f"no completed Harrier artifact directory for {source}")
-        if len(paths) != 1:
-            raise ValueError(f"expected one Harrier artifact directory for {source}, found {paths}")
-        resolved[source] = paths[0]
-    if not resolved:
-        raise ValueError(f"no completed Harrier artifacts under {EMBEDDING_ROOT}")
-    return resolved
-
-
-def _sample_shard(
-    batches: Iterator[list[dict[str, Any]]],
-    _shard: ShardInfo,
-    *,
-    source: str,
-    path: str,
-    rows: int,
-    quota: int,
-    seed: int,
-) -> Iterator[dict[str, Any]]:
-    digest = hashlib.sha256(f"{seed}:{source}:{path}".encode()).digest()
-    rng = np.random.default_rng(int.from_bytes(digest[:8], "little"))
-    selected = np.sort(rng.choice(rows, size=quota, replace=False))
-    offset = 0
-    written = 0
-    for batch in batches:
-        begin = int(np.searchsorted(selected, offset, side="left"))
-        end = int(np.searchsorted(selected, offset + len(batch), side="left"))
-        for index in selected[begin:end] - offset:
-            yield {"source": source, "embedding": batch[int(index)]["embedding"]}
-            written += 1
-        offset += len(batch)
-    if offset != rows or written != quota:
-        raise ValueError(f"sampled {written}/{quota} rows after reading {offset}/{rows} from {path}")
-    counters.pipeline.update_counter("harrier_cluster/sample_rows", written)
-
-
-def _sample_source(sample: SourceSample, output_path: str) -> None:
-    quotas = _part_quotas(sample.rows, sample.quota)
-    selected = [
-        (path, rows, quota) for path, rows, quota in zip(sample.paths, sample.rows, quotas, strict=True) if quota
-    ]
-    source_dir = prefix_join(output_path, sample.source.replace("/", "-"))
-    output_paths = tuple(prefix_join(source_dir, f"sample-{index:06d}.parquet") for index in range(len(selected)))
-    dataset = (
-        Dataset.from_list([InputFileSpec(path=path, columns=["embedding"]) for path, _, _ in selected])
-        .flat_map(load_file)
-        .window(4_096)
-        .map_shard(
-            lambda batches, shard, parts=tuple(selected), source=sample.source: _sample_shard(
-                batches,
-                shard,
-                source=source,
-                path=parts[shard.shard_idx][0],
-                rows=parts[shard.shard_idx][1],
-                quota=parts[shard.shard_idx][2],
-                seed=SEED,
-            )
-        )
-        .write_parquet(lambda shard_idx, _total: output_paths[shard_idx], schema=_SAMPLE_SCHEMA, skip_existing=True)
-    )
-    ZephyrContext(
-        resources=SAMPLE_WORKER_RESOURCES,
-        coordinator_resources=ResourceConfig.with_cpu(cpu=1, ram="8g", preemptible=False),
-        max_workers=min(MAX_WORKERS, len(selected)),
-        name=f"sample-harrier-{hashlib.sha256(sample.source.encode()).hexdigest()[:12]}",
-        stage_runner_factory=InlineRunner,
-    ).execute(dataset, verbose=True)
-
-
-def sample_embeddings(output_path: str, source_names: tuple[str, ...]) -> None:
-    embedding_paths = _completed_embedding_paths(source_names)
-    samples = _source_samples(embedding_paths)
-    with ThreadPoolExecutor(max_workers=PARALLEL_SOURCES) as pool:
-        futures = {pool.submit(_sample_source, sample, output_path): sample.source for sample in samples}
-        for future in as_completed(futures):
-            future.result()
-            logger.info("Sampled %s", futures[future])
-
-    source_rows = {sample.source: sum(sample.rows) for sample in samples}
-    source_quotas = {sample.source: sample.quota for sample in samples}
-    with open_url(prefix_join(output_path, "sample_stats.json"), "w") as file:
-        json.dump(
-            {
-                "target_rows": TARGET_ROWS,
-                "source_count": len(samples),
-                "source_rows": source_rows,
-                "source_quotas": source_quotas,
-                "embedding_paths": embedding_paths,
-                "small_source_max_rows": SMALL_SOURCE_MAX_ROWS,
-                "small_source_quota": SMALL_SOURCE_QUOTA,
-                "seed": SEED,
-            },
-            file,
-            indent=2,
-            sort_keys=True,
-        )
-
-
-def _stats_payload(value: ClusteringStats) -> dict[str, Any]:
-    return {
-        "loss": value.loss,
-        "weighted_mean_cosine_distance": value.weighted_mean_cosine_distance,
-        "minimum_weight_fraction": value.minimum_weight_fraction,
-        "maximum_weight_fraction": value.maximum_weight_fraction,
-        "weight_fractions": value.weight_fractions.tolist(),
-        "fine_centroid_counts": value.fine_centroid_counts.tolist(),
-    }
-
-
-def _coarsening_payload(result: CoarseningResult) -> dict[str, Any]:
-    return {
-        "method": "weighted_divisive_spherical_kmeans_then_single_centroid_hill_climb",
-        "minimum_fraction": COARSENING_CONFIG.minimum_fraction,
-        "seeds": list(COARSENING_CONFIG.seeds),
-        "split_restarts": COARSENING_CONFIG.split_restarts,
-        "split_iterations": COARSENING_CONFIG.split_iterations,
-        "selected_seed": result.selected_seed,
-        "divisive_runs": [{"seed": run.seed, "stats": _stats_payload(run.stats)} for run in result.divisive_runs],
-        "initial": _stats_payload(result.initial),
-        "final": _stats_payload(result.final),
-        "moves": result.moves,
-        "sweeps": result.sweeps,
-    }
-
-
-def _read_sample_table(path: str) -> pa.Table:
-    with StoragePath(path).open("rb") as file:
-        return pq.read_table(file, columns=["source", "embedding"])
-
-
-def _load_training_embeddings(
-    sample_path: str,
-    target: int,
-    seed: int,
-) -> _TrainingData:
-    paths = sorted(str(path) for path in StoragePath(prefix_join(sample_path, "**/*.parquet")).glob())
-    if not paths:
-        raise FileNotFoundError(f"No centroid sample shards under {sample_path}")
-
-    started = time.monotonic()
-    with ThreadPoolExecutor(max_workers=LOAD_PARALLELISM) as pool:
-        tables = list(pool.map(_read_sample_table, paths))
-    table = pa.concat_tables(tables)
-    encoded_sources = table["source"].combine_chunks().dictionary_encode()
-    source_names = tuple(encoded_sources.dictionary.to_pylist())
-    source_codes = encoded_sources.indices.to_numpy(zero_copy_only=False)
-    selected, quotas = _stratified_indices(source_codes, source_names, target, seed)
-
-    embeddings_column = table["embedding"].combine_chunks()
-    if not pa.types.is_fixed_size_list(embeddings_column.type) or embeddings_column.type.list_size != HARRIER_DIM:
-        raise ValueError(f"Unexpected centroid sample type {embeddings_column.type}")
-    quantized = embeddings_column.values.to_numpy(zero_copy_only=False).reshape(-1, HARRIER_DIM)
-    training_embeddings = np.ascontiguousarray(dequantize_to_fp32(quantized[selected]))
-    logger.info(
-        "Loaded %d x %d population and selected %d stratified training rows in %.1fs",
-        len(quantized),
-        HARRIER_DIM,
-        len(training_embeddings),
-        time.monotonic() - started,
-    )
-    return _TrainingData(training_embeddings, quantized, quotas)
-
-
-def _cluster_weights(index: _SearchIndex, quantized: np.ndarray) -> np.ndarray:
-    weights = np.zeros(K_TRAIN, dtype=np.int64)
-    for start in range(0, len(quantized), ASSIGN_BATCH_ROWS):
-        batch = dequantize_to_fp32(quantized[start : start + ASSIGN_BATCH_ROWS])
-        _, assignments = index.search(batch, 1)
-        weights += np.bincount(assignments[:, 0], minlength=K_TRAIN)
-    return weights
-
-
-def _save_npy(array: np.ndarray, output_path: str, name: str) -> None:
-    with open_url(prefix_join(output_path, name), "wb") as file:
-        np.save(file, array)
-
-
-def _train_fine_centroids(embeddings: np.ndarray) -> _TrainedCentroids:
-    import faiss  # noqa: PLC0415
-
-    faiss.omp_set_num_threads(N_THREADS)
-    started = time.monotonic()
-    kmeans = faiss.Kmeans(
-        d=HARRIER_DIM,
-        k=K_TRAIN,
-        niter=N_ITER,
-        nredo=N_REDO,
-        spherical=True,
-        seed=SEED,
-        verbose=True,
-        max_points_per_centroid=TRAIN_POINTS_PER_CENTROID,
-    )
-    kmeans.train(embeddings)
-    return _TrainedCentroids(
-        centroids=kmeans.centroids.astype(np.float32, copy=False),
-        index=kmeans.index,
-        seconds=time.monotonic() - started,
-        objective=float(kmeans.obj[-1]),
-    )
-
-
-def _write_training_artifact(
-    output_path: str,
-    sample_path: str,
-    training: _TrainingData,
-    trained: _TrainedCentroids,
-    weights: np.ndarray,
-    coarsening: CoarseningResult,
-) -> None:
-    _save_npy(trained.centroids, output_path, f"centroids_{K_TRAIN}.npy")
-    _save_npy(coarsening.fine_to_coarse, output_path, f"lookup_{K_TRAIN}_to_{K_COARSE}.npy")
-    with open_url(prefix_join(output_path, "fine_cluster_weights.json"), "w") as file:
-        json.dump({"document_counts_5000": weights.astype(int).tolist()}, file)
-    with open_url(prefix_join(output_path, "train_stats.json"), "w") as file:
-        json.dump(
-            {
-                "sample_path": sample_path,
-                "embedding_dim": HARRIER_DIM,
-                "k_train": K_TRAIN,
-                "k_views": [K_COARSE],
-                "n_sample": len(training.embeddings),
-                "population_rows": len(training.population),
-                "training_points_per_centroid": TRAIN_POINTS_PER_CENTROID,
-                "training_rows_by_source": training.rows_by_source,
-                "n_iter": N_ITER,
-                "n_redo": N_REDO,
-                "seed": SEED,
-                "n_threads": N_THREADS,
-                "train_seconds": trained.seconds,
-                "final_objective": trained.objective,
-                "coarsening": _coarsening_payload(coarsening),
-            },
-            file,
-            indent=2,
-        )
-
-
-def train_centroids(output_path: str, sample_path: str) -> None:
-    from threadpoolctl import threadpool_limits  # noqa: PLC0415
-
-    training = _load_training_embeddings(sample_path, TRAIN_ROWS, SEED)
-    with threadpool_limits(limits=N_THREADS):
-        trained = _train_fine_centroids(training.embeddings)
-        weights = _cluster_weights(trained.index, training.population).astype(np.float64)
-        if np.any(weights == 0):
-            raise ValueError("K-means produced an empty fine cluster")
-        coarsening = coarsen_centroids(trained.centroids, weights, COARSENING_CONFIG)
-    _write_training_artifact(output_path, sample_path, training, trained, weights, coarsening)
-
 
 def build_steps() -> tuple[StepSpec, StepSpec]:
-    source_names = tuple(sorted(select_sources()))
+    embedding_steps = build_harrier_embeddings("unused")
+    embedding_paths = {step.name.removeprefix("datakit/embed/harrier/"): step.output_path for step in embedding_steps}
     sample_step = StepSpec(
         name="datakit/cluster/domain/v1/harrier-all-sources-10m/sample",
-        deps=[],
+        deps=embedding_steps,
         hash_attrs={
-            "embedding_root": EMBEDDING_ROOT,
-            "source_names": source_names,
             "target_rows": TARGET_ROWS,
             "small_source_max_rows": SMALL_SOURCE_MAX_ROWS,
             "small_source_quota": SMALL_SOURCE_QUOTA,
             "seed": SEED,
-            "v": 4,
+            "version": PIPELINE_VERSION,
         },
         fn=remote(
-            lambda output_path, sources=source_names: sample_embeddings(output_path, sources),
+            lambda output_path, paths=embedding_paths: sample_centroid_inputs(
+                output_path=output_path,
+                embedding_paths=paths,
+                target_rows=TARGET_ROWS,
+                small_source_max_rows=SMALL_SOURCE_MAX_ROWS,
+                small_source_quota=SMALL_SOURCE_QUOTA,
+                seed=SEED,
+                worker_resources=SAMPLE_WORKER_RESOURCES,
+                coordinator_resources=SAMPLE_COORDINATOR_RESOURCES,
+                max_workers=MAX_WORKERS,
+                parallel_sources=PARALLEL_SOURCES,
+                load_parallelism=LOAD_PARALLELISM,
+            ),
             resources=DRIVER_RESOURCES,
             pip_dependency_groups=["datakit"],
         ),
@@ -528,10 +106,23 @@ def build_steps() -> tuple[StepSpec, StepSpec]:
             "seed": SEED,
             "n_threads": N_THREADS,
             "coarsening": COARSENING_CACHE_KEY,
-            "v": 2,
+            "version": PIPELINE_VERSION,
         },
         fn=remote(
-            lambda output_path, sample_path=sample_step.output_path: train_centroids(output_path, sample_path),
+            lambda output_path, sample_path=sample_step.output_path: train_centroids(
+                output_path=output_path,
+                sample_path=sample_path,
+                k_train=K_TRAIN,
+                train_rows=TRAIN_ROWS,
+                points_per_centroid=TRAIN_POINTS_PER_CENTROID,
+                n_iter=N_ITER,
+                n_redo=N_REDO,
+                seed=SEED,
+                n_threads=N_THREADS,
+                assign_batch_rows=ASSIGN_BATCH_ROWS,
+                load_parallelism=LOAD_PARALLELISM,
+                coarsening_config=COARSENING_CONFIG,
+            ),
             resources=TRAIN_RESOURCES,
             pip_dependency_groups=["datakit"],
         ),
